# `torch.compile` compatibility with `generate` + static cache