This PR adds the support of `torch.compile` when using `generate` + static cache.