@fxmarty 2024-02-20
@ArthurZucker @gante @LysandreJik This PR fixes many issues with the current `torch.compile` + static cache + generate implementation, as follow:

## 1. Always keep the same stride for inputs in the decode phase

`generate` apparently does not use directly the [`next_tokens`](https://github.com/huggingface/transformers/blob/1c81132e80478e278681686fe44dfec793d5dee9/src/transformers/generation/utils.py#L2440) variable as the next `input_ids`. Instead, the `next_tokens` are concatenated with previous tokens, and then [sliced](https://github.com/huggingface/transformers/blob/1c81132e80478e278681686fe44dfec793d5dee9/src/transformers/models/llama/modeling_llama.py#L1266), which results in the input tensors having different stride while having the same shape:

```
------------- loop forward in generate 0
model_inputs input ids shape torch.Size([2, 7])
model_inputs input ids stride (7, 1)
model_inputs position_ids shape torch.Size([2, 7])
model_inputs position_ids stride (7, 1)
------------- loop forward in generate 1
model_inputs input ids shape torch.Size([2, 1])
model_inputs input ids stride (8, 1)
model_inputs position_ids shape torch.Size([2, 1])
model_inputs position_ids stride (8, 1)
------------- loop forward in generate 2
model_inputs input ids shape torch.Size([2, 1])
model_inputs input ids stride (9, 1)
model_inputs position_ids shape torch.Size([2, 1])
model_inputs position_ids stride (9, 1)
------------- loop forward in generate 3
model_inputs input ids shape torch.Size([2, 1])
model_inputs input ids stride (10, 1)
model_inputs position_ids shape torch.Size([2, 1])
model_inputs position_ids stride (10, 1)
------------- loop forward in generate 4
etc.
```

This is bad because with torch.compile there are guards on the stride of the inputs, and thus recompilation is triggered in the decode phase while this is really not necessary.

```
V0220 17:27:08.283680 140705660285312 torch/_dynamo/guards.py:1381] Recompiling function forward in /home/felix/transformers/src/transformers/models/llama/modeling_llama.py:1127
V0220 17:27:08.283680 140705660285312 torch/_dynamo/guards.py:1381]     triggered by the following guard failure(s):
V0220 17:27:08.283680 140705660285312 torch/_dynamo/guards.py:1381]     - tensor 'L['input_ids']' stride mismatch at index 0. expected 7, actual 8

```
&
```
V0220 17:27:12.636874 140705660285312 torch/_dynamo/guards.py:1381] Recompiling function torch_dynamo_resume_in_forward_at_989 in /home/felix/transformers/src/transformers/models/llama/modeling_llama.py:989
V0220 17:27:12.636874 140705660285312 torch/_dynamo/guards.py:1381]     triggered by the following guard failure(s):
V0220 17:27:12.636874 140705660285312 torch/_dynamo/guards.py:1381]     - tensor 'L['position_ids']' stride mismatch at index 0. expected 7, actual 8
```

## 2. Do not compile `_update_causal_mask`

[`_update_causal_mask`](https://github.com/huggingface/transformers/blob/1c81132e80478e278681686fe44dfec793d5dee9/src/transformers/models/llama/modeling_llama.py#L1081) uses the input `attention_mask` length in its code.  I believe this results in an FX `placehoder` being a SymInt,

```
V0220 11:30:30.341809 140023118176640 torch/_dynamo/output_graph.py:1084] [2/1]  ===== __compiled_fn_12 =====
V0220 11:30:30.341809 140023118176640 torch/_dynamo/output_graph.py:1084] [2/1]  /home/felix/miniconda3/envs/fx/lib/python3.9/site-packages/torch/fx/_lazy_graph_module.py class GraphModule(torch.nn.Module):
V0220 11:30:30.341809 140023118176640 torch/_dynamo/output_graph.py:1084] [2/1]     def forward(self, s0 : torch.SymInt, L_attention_mask_ : torch.Tensor):
V0220 11:30:30.341809 140023118176640 torch/_dynamo/output_graph.py:1084] [2/1]         l_attention_mask_ = L_attention_mask_
```

which retriggers CUDAGraph capture for every decode step:
```
I0220 11:30:57.881897 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 13
I0220 11:30:57.902060 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 14
I0220 11:30:57.922157 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 15
I0220 11:30:57.942382 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 16
I0220 11:30:57.962995 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 17
I0220 11:30:57.983414 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 18
I0220 11:30:58.004108 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 19
I0220 11:30:58.024602 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 20
I0220 11:30:58.045034 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 21
I0220 11:30:58.065743 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 22
I0220 11:30:58.086160 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 23
I0220 11:30:58.106957 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 24
I0220 11:30:58.127666 140023118176640 torch/_inductor/cudagraph_trees.py:375] recording cudagraph tree for symint key 25
```

This is very slow. We avoid capturing `_update_causal_mask` with `@torch.compiler.disable`, which fixes the issue (no more cuda graph capture after the very first decode step).

## 3. Workaround for torch.compile bug not updating stateful `int`

On main, `StaticCache`'s [`seen_tokens`](https://github.com/huggingface/transformers/blob/main/src/transformers/cache_utils.py#L394) is bugged **only when using torch.compile** prior. Convince yourself with:

```python
from transformers import AutoTokenizer, AutoModelForCausalLM, GenerationConfig
import torch
from transformers.cache_utils import StaticCache

tokenizer = AutoTokenizer.from_pretrained(
    "NousResearch/Llama-2-7b-chat-hf", padding_side="left", pad_token="<s>"
)

with torch.device("cuda"):
    model = AutoModelForCausalLM.from_pretrained(
        "NousResearch/Llama-2-7b-chat-hf",
        torch_dtype=torch.float16,
        attn_implementation="sdpa",
    )

inputs = tokenizer(
    ["I would", "Today I am in Paris and"], padding=True, return_tensors="pt"
).to(model.device)

new_tokens = 10
gen_config = GenerationConfig(
    max_new_tokens=new_tokens,
    min_new_tokens=new_tokens,
    use_cache=True,
    pad_token_id=tokenizer.pad_token_id,
    num_beams=1,
    do_sample=False,
    eos_token_id=None,  # This is required for min_new_tokens to actually have an effect.
)
model.generation_config.eos_token_id = None  # greedy_search falls back on this eos_token_id that we need to set to None as well for min_new_tokens to have an effect.

gen_out = model.generate(**inputs, generation_config=gen_config)

decoded = tokenizer.batch_decode(gen_out, skip_special_tokens=True)

print("decoded", decoded)

print("compiling...")

model.forward = torch.compile(model.forward, mode="reduce-overhead")
print("Finished compile call")

# warmup
gen_out = model.generate(**inputs, generation_config=gen_config, cache_implementation="static")

print("\n\n\n\n\n\n----- second call")
gen_out = model.generate(**inputs, generation_config=gen_config, cache_implementation="static")

decoded = tokenizer.batch_decode(gen_out, skip_special_tokens=True)

print("decoded static", decoded)
```

which yields

![image](https://github.com/huggingface/transformers/assets/9808326/0fc8f657-6ae8-4e41-8fac-dc7a16fa9811)


Using a `torch.Tensor` updated in-place instead of `int` fixes the bug.

## 4. Workaround for torch.compile bug where subclasses are added after the `torch.compile` call

This is the case in `_setup_cache`. Even with the above fix, there is still a bug where `seen_tokens` is not properly updated. By making sure `_setup_cache` is called **BEFORE** `torch.compile`, the bug disappears.

Hence the API change and the introduction of `prepare_static_cache`.

I'll see if we can make it work WITHOUT an API change, completely removing `seen_tokens` from the modeling, or if that will still be an issue.

## Results

On `main`:

```
-------------- STATIC CACHE
compiling...
torch.compile call: 660.706 ms
/home/felix/miniconda3/envs/fx/lib/python3.9/site-packages/torch/_inductor/compile_fx.py:148: UserWarning: TensorFloat32 tensor cores for float32 matrix multiplication available but not enabled. Consider setting `torch.set_float32_matmul_precision('high')` for better performance.
  warnings.warn(

- 0-th `generate` call latency per token: 28623.272 ms

- 1-th `generate` call latency per token: 2767.040 ms

- 2-th `generate` call latency per token: 449.126 ms

- 3-th `generate` call latency per token: 11.163 ms

- 4-th `generate` call latency per token: 11.141 ms
decoded static ['I would like to know how to', 'Today I am in Paris and I am feeling very grateful']
```

On this branch:
```
-------------- STATIC CACHE
compiling...
torch.compile call: 690.987 ms
/home/felix/miniconda3/envs/fx/lib/python3.9/site-packages/torch/_inductor/compile_fx.py:148: UserWarning: TensorFloat32 tensor cores for float32 matrix multiplication available but not enabled. Consider setting `torch.set_float32_matmul_precision('high')` for better performance.
  warnings.warn(

- 0-th `generate` call latency per token: 17843.997 ms

- 1-th `generate` call latency per token: 591.632 ms

- 2-th `generate` call latency per token: 11.952 ms

- 3-th `generate` call latency per token: 11.924 ms

- 4-th `generate` call latency per token: 11.921 ms
decoded static ['I would like to know how to', 'Today I am in Paris and I am feeling very grateful']
```